Barretone bot fix
fixes #478
I could use the registry and get the hardness for the oak log; but since this is a fix related to baritone not sending enough data, I'm assuming a worst-case scenario where the vanilla function has been replaced with one that needs that missing data to. So I'm returning a constant so that we don't have a maybe.